Krylov operators

KrylovPreconditioners.KrylovOperator — Function

KrylovOperator(A; nrhs::Int=1, transa::Char='N')

Create a Krylov operator to accelerate sparse matrix-vector or matrix-matrix products on GPU architectures. The operator is compatible with sparse matrices stored on NVIDIA, AMD, and Intel GPUs.

Input arguments

A: The sparse matrix on the GPU that serves as the operator for matrix-vector or matrix-matrix products;

nrhs: Specifies the number of columns for the right-hand sides. Defaults to 1 for standard matrix-vector products;

transa: Determines how the matrix A is applied during the products; 'N' for no transposition, 'T' for transpose, and 'C' for conjugate transpose.

Output argument

op: An instance of AbstractKrylovOperator representing the Krylov operator for the specified sparse matrix and parameters.

KrylovPreconditioners.update! — Method

update!(op::AbstractKrylovOperator, A)

Update the sparse matrix A associated with the given AbstractKrylovOperator without the need to reallocate buffers or repeat the structural analysis phase for detecting parallelism for sparse matrix-vector or matrix-matrix products. A and the operator op must have the same sparsity pattern, enabling efficient reuse of existing resources.

Input arguments

op: The Krylov operator to update;

A: The new sparse matrix to associate with the operator.

## Nvidia GPUs

Sparse matrices have a specific storage on Nvidia GPUs (`CuSparseMatrixCSC`, `CuSparseMatrixCSR` or `CuSparseMatrixCOO`):

julia using CUDA, CUDA.CUSPARSE using SparseArrays using KrylovPreconditioners

if CUDA.functional()

CPU Arrays

A_cpu = sprand(200, 100, 0.3)

GPU Arrays

Acscgpu = CuSparseMatrixCSC(Acpu) Acsrgpu = CuSparseMatrixCSR(Acpu) Acoogpu = CuSparseMatrixCOO(A_cpu)

Krylov operators

opcsc = KrylovOperator(Acscgpu; nrhs=1, transa='N') opcsr = KrylovOperator(Acsrgpu; nrhs=1, transa='T') opcoo = KrylovOperator(Acoo_gpu; nrhs=5, transa='N') end

## AMD GPUs

Sparse matrices have a specific storage on AMD GPUs (`ROCSparseMatrixCSC`, `ROCSparseMatrixCSR` or `ROCSparseMatrixCOO`):

julia using AMDGPU, AMDGPU.rocSPARSE using SparseArrays using KrylovPreconditioners

if AMDGPU.functional()

CPU Arrays

A_cpu = sprand(200, 100, 0.3)

GPU Arrays

Acscgpu = ROCSparseMatrixCSC(Acpu) Acsrgpu = ROCSparseMatrixCSR(Acpu) Acoogpu = ROCSparseMatrixCOO(A_cpu)

Krylov operators

opcsc = KrylovOperator(Acscgpu; nrhs=1, transa='N') opcsr = KrylovOperator(Acsrgpu; nrhs=1, transa='T') opcoo = KrylovOperator(Acoo_gpu; nrhs=5, transa='N') end

## Intel GPUs

Sparse matrices have a specific storage on Intel GPUs (`oneSparseMatrixCSR`):

julia using oneAPI, oneAPI.oneMKL using SparseArrays using KrylovPreconditioners

if oneAPI.functional()

CPU Arrays

A_cpu = sprand(Float32, 20, 10, 0.3)

GPU Arrays

Acsrgpu = oneSparseMatrixCSR(A_cpu)

Krylov operator

opcsr = KrylovOperator(Acsr_gpu; nrhs=1, transa='N') end ```