Reference

Index

KrylovPreconditioners.BlockJacobiPreconditioner

KrylovPreconditioners._fillblock_gpu!

KrylovPreconditioners.build_adjmatrix

KrylovPreconditioners.overlap

KrylovPreconditioners.update!

KrylovPreconditioners._fillblock_gpu! — Method

_fillblock_gpu

Fill the dense blocks of the preconditioner from the sparse CSR matrix arrays

KrylovPreconditioners.build_adjmatrix — Method

build_adjmatrix

Build the adjacency matrix of a matrix A corresponding to the undirected graph

KrylovPreconditioners.overlap — Method

overlap(Graph, subset, level)

Given subset embedded within Graph, compute subset2 such that subset2 contains subset and all of its adjacent vertices.

KrylovPreconditioners.update! — Method

function update!(p, J::SparseMatrixCSC)

Update the preconditioner p from the sparse Jacobian J in CSC format for the CPU

Note that this implements the same algorithm as for the GPU and becomes very slow on CPU with growing number of blocks.

KrylovPreconditioners.BlockJacobiPreconditioner — Type

BlockJacobiPreconditioner

Overlapping-Schwarz preconditioner.

Attributes

nblocks::Int64: Number of partitions or blocks.

blocksize::Int64: Size of each block.

partitions::Vector{Vector{Int64}}:npart` partitions stored as lists

cupartitions: partitions transfered to the GPU

lpartitions::Vector{Int64}`: Length of each partitions.

culpartitions::Vector{Int64}`: Length of each partitions, on the GPU.

blocks: Dense blocks of the block-Jacobi

cublocks: Js transfered to the GPU

map: The partitions as a mapping to construct views

cumap: cumap transferred to the GPU`

part: Partitioning as output by Metis

cupart: part transferred to the GPU